Removed Engine component by mergin code with evaluator. Moved engine tests to evaluator component
